feat: got posts.php page working finally, thank goodness

+++ BloggingPlatform/modules/post/posts.php
@@ -10,8 +10,11 @@ require('../../config.php');
 require(DIR_HELPERS . 'session.php');
 require(DIR_HELPERS . 'db.php');
 
-if(isset($_GET['action']) & isset($_GET['post'])) {
-    $conn = dbConnect();
+$conn = dbConnect();
+
+
+if(isset($_GET['action']) && isset($_GET['post'])) {
+
     $post = $_GET['post'];
     $deleteQuery = "DELETE FROM post WHERE ID = $post";
     if($conn->query($deleteQuery) != true) {
@@ -28,12 +31,27 @@ $userId = getUserId();
 <script src="../helpers/languageFilter.js"></script>
 <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
 <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
 </head>
 <body>
 <?php include(VIEW_HEADER); ?>
 <h3>Your Posts</h3>
-<?php include '_getAllPosts.php'; ?>
+
+<?php
+
+$sqlQuery = "SELECT post.ID,post.Title,post.Content,CONCAT(user.FirstName, ' ',user.LastName) AS Author FROM post INNER JOIN user WHERE user.ID = post.Author";
+
+$resultSet = $conn->query($sqlQuery);
+
+echo "<ul style='margin-top:50px'>";
+while($row = $resultSet->fetch_assoc()) {
+    $id = $row["ID"];
+    $title = $row["Title"];
+    $author = $row["Author"];
+    $votes = $row["Votes"];
+    echo "<li><a href='viewPost.php?id=$id'>$title - By: $author, Votes: $votes</a></li>";
+}
+echo "</ul>";
+?>
 
 <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
